// packages/react-ui/components/FxInput/FxInput.d.ts
import * as React from 'react';

import { InputProps } from '../Input';

export type FxInputProps = InputProps & {
  auto?: boolean;
  type?: 'currency' | 'text';
  onRestore: () => void;
};

export interface FxInputState {}

export default class FxInput extends React.Component<
  FxInputProps,
  FxInputState
> {}